refactor(#2609): Rename dir to types
