+++ ui/src/app/dashboard/components/widgets/table/table-widget.component.ts
@@ -21,10 +21,10 @@ import { BaseStreamPipesWidget } from '../base/base-widget';
 import { StaticPropertyExtractor } from '../../../sdk/extractor/static-property-extractor';
 import { MatTableDataSource } from '@angular/material/table';
 import { TableConfig } from './table-config';
-import { SemanticTypeService } from '../../../../core-services/semantic-type/semantic-type.service';
 import { ResizeService } from '../../../services/resize.service';
 import { DatalakeRestService } from '@streampipes/platform-services';
 import { WidgetConfigBuilder } from '../../../registry/widget-config-builder';
+import { SemanticTypeService } from '../../../../core-services/types/semantic-type.service';
 
 @Component({
     selector: 'sp-table-widget',
